+++ src/main/java/com/comeon/backend/common/response/SliceResponse.java
@@ -1,4 +1,4 @@
-package com.comeon.backend.common.api;
+package com.comeon.backend.common.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
