+++ src/main/java/com/comeon/backend/image/infrastructure/S3FileManager.java
@@ -5,8 +5,8 @@ import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.comeon.backend.common.exception.CommonErrorCode;
-import com.comeon.backend.common.exception.RestApiException;
+import com.comeon.backend.common.error.CommonErrorCode;
+import com.comeon.backend.common.error.RestApiException;
 import com.comeon.backend.image.ImageErrorCode;
 import com.comeon.backend.image.application.FileManager;
 import com.comeon.backend.image.application.ImageDto;
