+++ src/main/java/com/comeon/backend/jwt/infra/JwtParserImpl.java
@@ -3,8 +3,8 @@ package com.comeon.backend.jwt.infra;
 import com.comeon.backend.jwt.application.TokenType;
 import com.comeon.backend.jwt.domain.JwtClaims;
 import com.comeon.backend.jwt.domain.JwtParser;
-import com.comeon.backend.common.exception.CommonErrorCode;
-import com.comeon.backend.common.exception.RestApiException;
+import com.comeon.backend.common.error.CommonErrorCode;
+import com.comeon.backend.common.error.RestApiException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
