+++ src/main/java/com/comeon/backend/meeting/presentation/NotMemberException.java
@@ -1,7 +1,7 @@
 package com.comeon.backend.meeting.presentation;
 
-import com.comeon.backend.common.exception.ErrorCode;
-import com.comeon.backend.common.exception.RestApiException;
+import com.comeon.backend.common.error.ErrorCode;
+import com.comeon.backend.common.error.RestApiException;
 import com.comeon.backend.meeting.MeetingErrorCode;
 
 public class NotMemberException extends RestApiException {
