+++ src/main/java/com/comeon/backend/meeting/presentation/api/MeetingExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.comeon.backend.meeting.presentation.api;
 
-import com.comeon.backend.common.api.ErrorResponse;
-import com.comeon.backend.common.util.ResponseEntityUtils;
+import com.comeon.backend.common.response.ErrorResponse;
+import com.comeon.backend.common.response.ResponseEntityUtils;
 import com.comeon.backend.meeting.presentation.MemberAuthorizationException;
 import com.comeon.backend.meeting.presentation.NotMemberException;
 import lombok.extern.slf4j.Slf4j;
