+++ src/test/java/com/comeon/backend/api/error/controller/ErrorRestDocsController.java
@@ -1,8 +1,8 @@
 package com.comeon.backend.api.error.controller;
 
-import com.comeon.backend.common.api.ErrorResponse;
-import com.comeon.backend.common.exception.CommonErrorCode;
-import com.comeon.backend.common.exception.ErrorCode;
+import com.comeon.backend.common.response.ErrorResponse;
+import com.comeon.backend.common.error.CommonErrorCode;
+import com.comeon.backend.common.error.ErrorCode;
 import com.comeon.backend.image.ImageErrorCode;
 import com.comeon.backend.jwt.JwtErrorCode;
 import com.comeon.backend.meeting.MeetingErrorCode;
